+++ test.py
@@ -1,10 +1,8 @@
-import classifiers as cf
-import cuts
-import binDynamic as bd
-import binStatic as bs
 from pandas import read_csv, np
-import voteSelection as vs
 import util as ut
+import cuts
+import binSelection as bs
+import classifiers as cf
 
 def artificialTest():
 	#files = ['data-f1.csv','data-f2.csv','data-f3.csv','data-f4.csv','simpleCorrelations.csv','data-r50.csv','data2400-r20.csv']
@@ -33,18 +31,18 @@ def artificialTest():
 		ss.append(bs.binStatic(data,1))
 		#ss.append(bs.binStatic(data,2))
 		#ss.append(bs.binStatic(data,3))
-		weight = (vs.sumMixedCorrelation(ss))
+		weight = (bs.sumMixedCorrelation(ss))
 		#print "static:",weight
 
 		#Dynamic search
 		cr = []
 		#binarySearchBinsX, y, method, split, useSteps, normalizeResult, Debug
-		cr.append(bd.binarySearchBins(X,y,0,0,2))
-		#cr.append(bd.binarySearchBins(X,y,1,0,2))
-		#cr.append(bd.binarySearchBins(X,y,2,0,2))
-		#cr = (vs.sumMixedCorrelation(cr))
+		#cr.append(bs.binarySearchBins(X,y,0,0,2))
+		#cr.append(bs.binarySearchBins(X,y,1,0,2))
+		#cr.append(bs.binarySearchBins(X,y,2,0,2))
+		#cr = (bs.sumMixedCorrelation(cr))
 		#print "dynamic:",cr
-		#bx = bd.cuadratureSearchBins(X)	
+		#bx = bs.cuadratureSearchBins(X)	
 		#print bx
 		
 		########### Cuts ###########
